refactor invincible powerup
