--- TheAlienThatEatsTheCarrot/GameEngine/Entities/Factories/PowerupFactories/InvinsiblePowerupFactory.swift
@@ -1,21 +1,0 @@
-//
-//  InvinsiblePowerupFactory.swift
-//  TheAlienThatEatsTheCarrot
-//
-//  Created by Justin Cheah Yun Fei on 29/3/24.
-//
-
-import Foundation
-
-class InvinsiblePowerupFactory: PowerupFactory {
-    override func createComponents() -> [Component] {
-        let size = CGSize(width: boardObject.width, height: boardObject.height)
-        let powerupComponent = PowerupComponent(entity: entity,
-                                                powerupType: .invinsible)
-        let renderableComponent = RenderableComponent(entity: entity,
-                                                      position: boardObject.position,
-                                                      objectType: .powerup(.invinsible),
-                                                      size: size)
-        return [powerupComponent, renderableComponent]
-    }
-}
